requirements.txt
fastapi
uvicorn[standard]
email-validator
dnspython
aiosmtplib
pydantic
rapidfuzz
